Add 'optional: true' to Piece model
The bug in the seeder was caused by seeded pieces failing to save when
they weren't bought, because the model was requiring user_id foreign
keys as a necessary field. The above change makes that field optional. I
also added pry to the Gemfile to debug this and will keep it for future
debugging.