Roll back
I've deleted the custom route and signup page for artists because that
requires me to create custom changes to the Devise controller, and that
has been too difficult. For now, I will try to simply conditionally add
additional information on the generic user signup page, and we can
change this later if we have time and if we decide it would be a good
idea.